Упорядочивает методы внутри классов, вносит другие мелкие правки для соответствия критериям
